+++ mex/editor/layout.py
@@ -1,31 +1,30 @@
-from collections import defaultdict
-from typing import Literal
-
 import reflex as rx
 
-from mex.editor.state import State
+from mex.editor.state import NavItem, State
 
-Header = Literal["search"] | Literal["edit"] | Literal["merge"]
+
+def user_button() -> rx.Component:
+    """Return a user button with an icon that indicates their access rights."""
+    return rx.button(
+        rx.cond(
+            State.user.write_access,  # type: ignore[union-attr]
+            rx.icon(tag="user_round_cog"),
+            rx.icon(tag="user_round"),
+        ),
+        variant="ghost",
+        style={"marginTop": "0"},
+    )
 
 
 def user_menu() -> rx.Component:
-    """Return a user menu button with an icon indicating their access rights."""
+    """Return a user menu with a trigger, the user's name and a logout button."""
     return rx.menu.root(
-        rx.menu.trigger(
-            rx.button(
-                rx.cond(
-                    State.user.write_access,  # type: ignore[union-attr]
-                    rx.icon(tag="user_round_cog"),
-                    rx.icon(tag="user_round"),
-                ),
-                variant="ghost",
-                style={"marginTop": "0"},
-            ),
-        ),
+        rx.menu.trigger(user_button()),
         rx.menu.content(
             rx.menu.item(
                 State.user.name,  # type: ignore[union-attr]
                 disabled=True,
+                style={"color": "var(--gray-12)"},
             ),
             rx.menu.separator(),
             rx.menu.item(
@@ -33,41 +32,41 @@ def user_menu() -> rx.Component:
                 on_select=State.logout,
             ),
         ),
+        custom_attrs={"data-testid": "user-menu"},
     )
 
 
-def nav_links(heading: Header) -> list[rx.Component]:
-    """Return a list of navigation links with the current heading highlighted."""
-    should_underline = defaultdict(lambda: "none", {heading: "always"})
-    return [
-        rx.link(
-            "Search",
-            href="/",
-            underline=should_underline["search"],
-            padding="0.2em",
-        ),
-        rx.link(
-            "Edit",
-            href="/item/123",
-            underline=should_underline["edit"],
-            padding="0.2em",
-        ),
-        rx.link(
-            "Merge",
-            href="/merge",
-            underline=should_underline["merge"],
-            padding="0.2em",
-        ),
-    ]
+def nav_link(item: NavItem) -> rx.Component:
+    """Return a link component for the given navigation item."""
+    return rx.link(
+        item.title,
+        href=item.href,
+        underline=item.underline,
+        padding="0.2em",
+    )
 
 
-def nav_bar(heading: Header) -> rx.Component:
+def mex_editor_logo() -> rx.Component:
+    """Return the editor's logo with icon and label."""
+    return rx.hstack(
+        rx.icon(
+            "file-pen-line",
+            size=28,
+        ),
+        rx.heading(
+            "MEx Editor",
+            weight="medium",
+            style={"user-select": "none"},
+        ),
+    )
+
+
+def nav_bar() -> rx.Component:
     """Return a navigation bar component."""
     return rx.hstack(
-        rx.icon("file-pen-line", size=28),
-        rx.heading("MEx Editor", weight="medium"),
+        mex_editor_logo(),
         rx.divider(orientation="vertical", size="2"),
-        *nav_links(heading),
+        rx.foreach(State.nav_items, nav_link),
         rx.divider(orientation="vertical", size="2"),
         user_menu(),
         rx.spacer(),
@@ -75,22 +74,31 @@ def nav_bar(heading: Header) -> rx.Component:
         id="navbar",
         padding="1em",
         position="fixed",
-        style=rx.style.Style(background="var(--accent-4)"),
+        style={"background": "var(--accent-4)"},
         top="0px",
         width="100%",
         z_index="1000",
+        custom_attrs={"data-testid": "nav-bar"},
     )
 
 
-def page(heading: Header, *children: str | rx.Component) -> rx.Component:
+def page(*children: str | rx.Component) -> rx.Component:
     """Return a page fragment with navigation bar and given children."""
-    return rx.fragment(
-        nav_bar(heading),
-        rx.vstack(
-            *children,
-            justify="center",
-            min_height="85vh",
-            margin="4em 1em 1em",
-            spacing="5",
+    return rx.cond(
+        State.user,
+        rx.fragment(
+            nav_bar(),
+            rx.vstack(
+                *children,
+                min_height="85vh",
+                margin="2em 1em 1em",
+                spacing="5",
+            ),
+            on_mount=State.load_page,
+            on_unmount=State.unload_page,
+        ),
+        rx.center(
+            rx.spinner(size="3"),
+            style={"marginTop": "40vh"},
         ),
     )